Added KeySwitchKey generation

+++ pyFHE/trgsw.py
@@ -18,4 +18,4 @@ def trgswSymEncrypt(p, alpha, h, key, twist):
 
 def trgswExternalProdcut(g,r,params):
     decvec = Decomposition(r,params)
-    return np.array([np.sum([PolyMul(decvec[i], g[i][0],params.twist) for i in range(2 * params.l)],axis = 0),np.sum([PolyMul(decvec[i], g[i][1],params.twist) for i in range(2 * params.l)], axis = 0)])
+    return np.array([np.sum([PolyMul(decvec[i], g[i][0],params.twist) for i in range(2 * params.l)],axis = 0),np.sum([PolyMul(decvec[i], g[i][1],params.twist) for i in range(2 * params.l)], axis = 0)],dtype = np.uint32)
